+++ apps/docs/gatsby-config.js
@@ -1,65 +1,87 @@
-const { cleanDoclets } = require('gatsby-transformer-react-docgen/doclets');
-const path = require('path');
-const remarkSlug = require('remark-slug');
-const defaultDescriptions = require('./src/defaultPropDescriptions');
-const algoliaSearch = require('./algoliaSearch');
-const ahaReactConfig = require('./config');
+const { cleanDoclets } = require("gatsby-transformer-react-docgen/doclets");
+const path = require("path");
+const remarkSlug = require("remark-slug");
+const defaultDescriptions = require("./src/defaultPropDescriptions");
+const algoliaSearch = require("./algoliaSearch");
+const ahaReactConfig = require("./config");
 
-const activeEnv = process.env.ENV || 'dev';
+const activeEnv = process.env.ENV || "dev";
 console.log(`Using environment config: '${activeEnv}'`);
-require('dotenv').config({
+require("dotenv").config({
   path: `.env.${activeEnv}`,
 });
 
 module.exports = {
   siteMetadata: {
-    title: 'Aha Design System - Documentation',
-    description: 'An ever-evolving system that enables us to build higher quality products more efficiently',
-    author: 'GotIt, Inc. contributors',
+    title: "Aha Design System - Documentation",
+    description:
+      "An ever-evolving system that enables us to build higher quality products more efficiently",
+    author: "GotIt, Inc. contributors",
     browsers: [
-      'last 4 Chrome versions',
-      'last 4 Firefox versions',
-      'last 2 Edge versions',
-      'last 2 Safari versions',
+      "last 4 Chrome versions",
+      "last 4 Firefox versions",
+      "last 2 Edge versions",
+      "last 2 Safari versions",
     ],
   },
-  pathPrefix: 'aha/' + ahaReactConfig.version,
+  flags: {
+    PRESERVE_FILE_DOWNLOAD_CACHE: true,
+    PARALLEL_SOURCING: true,
+    FAST_DEV: true,
+  },
+  pathPrefix: ahaReactConfig.version,
   plugins: [
-    'gatsby-plugin-sorted-assets',
+    "gatsby-plugin-sorted-assets",
     {
-      resolve: 'gatsby-plugin-mdx',
+      resolve: "gatsby-plugin-mdx",
       options: {
         defaultLayouts: {
-          default: require.resolve('./src/layouts/ApiLayout'),
+          default: require.resolve("./src/layouts/ApiLayout"),
         },
         remarkPlugins: [remarkSlug],
       },
     },
     {
-      resolve: 'gatsby-source-filesystem',
+      resolve: "gatsby-source-filesystem",
       options: {
-        path: path.resolve(__dirname, '../../packages/react/src/'),
-        name: 'source',
+        path: path.resolve(__dirname, "../../packages/react/src/"),
+        name: "source",
       },
     },
     {
-      resolve: 'gatsby-source-filesystem',
+      resolve: "gatsby-plugin-root-import",
       options: {
-        path: path.resolve(__dirname, './src/pages'),
-        name: 'pages',
+        utils: path.resolve(__dirname, "../../packages/react/src/utils/"),
+        components: path.resolve(
+          __dirname,
+          "../../packages/react/src/components/"
+        ),
+        hooks: path.resolve(__dirname, "../../packages/react/src/hooks/"),
+        constants: path.resolve(
+          __dirname,
+          "../../packages/react/src/constants/"
+        ),
+        "react-dom": path.resolve(__dirname, "../../node_modules/react-dom"),
       },
     },
     {
-      resolve: 'gatsby-source-filesystem',
+      resolve: "gatsby-source-filesystem",
       options: {
-        path: path.resolve(__dirname, '../../packages/react/CHANGELOG.md'),
-        name: 'CHANGELOG',
+        path: path.resolve(__dirname, "./src/pages"),
+        name: "pages",
       },
     },
     {
-      resolve: 'gatsby-transformer-react-docgen',
+      resolve: "gatsby-source-filesystem",
       options: {
-        resolver: require('./resolveHocComponents'),
+        path: path.resolve(__dirname, "../../packages/react/CHANGELOG.md"),
+        name: "CHANGELOG",
+      },
+    },
+    {
+      resolve: "gatsby-transformer-react-docgen",
+      options: {
+        resolver: require("./resolveHocComponents"),
         handlers: [
           function defaultDescriptionsHandler(docs) {
             docs._props.forEach((_, name) => {
@@ -67,7 +89,9 @@ module.exports = {
                 const prop = docs.getPropDescriptor(name);
                 const dflt = defaultDescriptions[name];
 
-                if (dflt && !cleanDoclets(prop.description)) { prop.description = `${dflt}\n${prop.description}`; }
+                if (dflt && !cleanDoclets(prop.description)) {
+                  prop.description = `${dflt}\n${prop.description}`;
+                }
               }
             });
           },
@@ -75,34 +99,31 @@ module.exports = {
       },
     },
     {
-      resolve: 'gatsby-transformer-remark',
+      resolve: "gatsby-transformer-remark",
       options: {
         plugins: [
-          'gatsby-remark-prismjs',
+          "gatsby-remark-prismjs",
           {
-            resolve: 'gatsby-remark-external-links',
+            resolve: "gatsby-remark-external-links",
             options: {
-              target: '_blank',
-              rel: 'noopener noreferrer',
+              target: "_blank",
+              rel: "noopener noreferrer",
             },
-          }],
+          },
+        ],
       },
     },
-    'gatsby-plugin-catch-links',
-    'gatsby-plugin-sass',
+    "gatsby-plugin-catch-links",
+    "gatsby-plugin-sass",
     {
-      resolve: 'gatsby-plugin-astroturf',
-      options: { extension: '.module.scss' },
-    },
-    {
-      resolve: 'gatsby-plugin-nprogress',
+      resolve: "gatsby-plugin-nprogress",
       options: {
-        color: 'var(--colorPrimary)',
+        color: "var(--colorPrimary)",
         showSpinner: false,
       },
     },
     {
-      resolve: 'gatsby-plugin-algolia',
+      resolve: "gatsby-plugin-algolia",
       options: {
         appId: process.env.GATSBY_ALGOLIA_APP_ID,
         apiKey: process.env.GATSBY_ALGOLIA_API_KEY,