+++ packages/react/src/components/Calender/DatePicker.js
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import DatePickerBase from 'react-date-picker/dist/entry.nostyle';
+import DatePickerBase from 'react-date-picker';
 import Context from 'components/Form/Context';
 import Icon from 'components/Icon';
-import DatePickerBaseV2 from './v2/DatePicker/DatePicker';
+import DatePickerBaseV2 from 'components/Calender/v2/DatePicker/DatePicker';
 
 const DatePicker = React.forwardRef(({ className, noClearIcon, size, version, calendarClassName, ...props }, ref) => {
   const { sizeControl } = useContext(Context);