+++ packages/react/src/components/Tag/index.js
@@ -33,7 +33,7 @@ const propTypes = {
 };
 const defaultProps = {
   variant: 'primary',
-  textClassName: false,
+  textClassName: '',
 };