+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Release 2.0.0 - January 22, 2021
-* First release
+
+- First release
+
 ### Added
-* Icons: list of all icons with name and category
-* createIconAssetPlugin(): utils function to create asset plugin for Aha React Icon component
+
+- Icons: list of all icons with name and category
+- createIconAssetPlugin(): utils function to create asset plugin for Aha React Icon component